refactor JournalManager out of Application. Created OptionManager to handle SharedPreferences.
