+++ app/src/main/java/edu/uw/ischool/opensecrets/MainActivity.kt
@@ -26,14 +26,11 @@ class MainActivity : AppCompatActivity() {
         val journal = File(filesDir, "journal.json")
 
         Log.d("file", journal.path.toString())
-        val prefs =
-            this.getSharedPreferences(getString(R.string.preference_key), Context.MODE_PRIVATE)
-        Log.d("filePrefAll", prefs.all.toString())
-        if (prefs.getString("username", null) == null ||
-            prefs.getString("password", null) == null
+        if ((this.application as SecretApp).optionManager.getUsername() == null ||
+            (this.application as SecretApp).optionManager.getPassword() == null
         ) {
             Log.d("fileExist", journal.exists().toString())
-            if (!(this.application as SecretApp).journalExist()) {
+            if (!(this.application as SecretApp).journalManager.journalExist()) {
                 Log.d("fileCreate", journal.createNewFile().toString())
             }
             startActivity(Intent(this, LoginActivity::class.java))
@@ -51,7 +48,7 @@ class MainActivity : AppCompatActivity() {
             this.supportFragmentManager.setFragmentResultListener(
                 "delete_event",
                 this
-            ) { key, bundle ->
+            ) { _, bundle ->
 
                 val result = bundle.getString("delete_event").toBoolean()
                 if (result) {
@@ -92,7 +89,7 @@ class MainActivity : AppCompatActivity() {
             }
 
             // load data.
-            val entries = (this.application as SecretApp).loadEntry()
+            val entries = (this.application as SecretApp).journalManager.loadEntry()
 
             // check data before loading correct view.
             if (entries.isNullOrEmpty()) {
@@ -113,7 +110,7 @@ class MainActivity : AppCompatActivity() {
      */
     private fun deleteEntry(pos: Int) {
         val dialog =
-            DeleteEntryDialogFragment((this.application as SecretApp)::deleteEntry)
+            DeleteEntryDialogFragment((this.application as SecretApp).journalManager::deleteEntry)
         val args = Bundle()
         args.putInt("pos", pos)
         dialog.arguments = args
@@ -136,7 +133,7 @@ class MainActivity : AppCompatActivity() {
                 // Use the Builder class for convenient dialog construction.
                 val builder = AlertDialog.Builder(it)
                 builder.setMessage("Do you want to delete this entry?")
-                    .setPositiveButton("Yes") { dialog, id ->
+                    .setPositiveButton("Yes") { _, _ ->
                         val pos = arguments?.getInt("pos")
                         if (pos != null) {
                             if (deleteFn(pos)) {
@@ -153,7 +150,7 @@ class MainActivity : AppCompatActivity() {
 
                         }
                     }
-                    .setNegativeButton("No") { dialog, id ->
+                    .setNegativeButton("No") { _, _ ->
                         // User cancelled the dialog.
                     }
                 // Create the AlertDialog object and return it.
